Did a little bit of a rewrite to fix things up
The other release still utilized a lot of faulty and slow logic, everything works seamlessly now.

I still don't currently have the time to finish this project since I'm working on both AJB and Paragone, this is just a little bit of a rebase so that it's ready for me when I get back to it.